fix(llmisvc): skip migration swap for external InferencePool refs
The migration backendRef swap logic was incorrectly applying to external
InferencePool references (user-managed pools). This fix:

- Skips adjustBackendRefForMigration when Pool.HasRef() is true
- Adds BackendRefInferencePoolV1Alpha2 helper to fixture
- Updates controller_int_test to wait for migration before asserting v1 backendRef
